Move old org and api files to dynamodb and rest packages temporarily
